% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alpha_blood}
\alias{alpha_blood}
\title{Recommended values for a Dirichlet prior on cell composition in blood samples.
Estimated by fitting Dirichlet distribution to cell counts data.}
\format{A vector of length 6}
\usage{
alpha_blood
}
\description{
Recommended values for a Dirichlet prior on cell composition in blood samples.
Estimated by fitting Dirichlet distribution to cell counts data.
}
\keyword{datasets}
