% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{baseline_GSE40279}
\alias{baseline_GSE40279}
\title{Cell composition for 650 individuals from GSE40279. Estimated by running an array method on the original (array) data.}
\format{A data frame with 650 rows and 6 variables}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE40279}
}
\usage{
baseline_GSE40279
}
\description{
Cell composition for 650 individuals from GSE40279. Estimated by running an array method on the original (array) data.
}
\keyword{datasets}
