% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methylation_GSE40279}
\alias{methylation_GSE40279}
\title{Simulated amount of methylated reads for 650 individuals from GSE40279. The data was sub-sampled to simulate a 30X coverage.
Only 241 sites that are known to differ substaintally between cell types are recorded.}
\format{A data frame with 650 rows and 241 variables}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE40279}
}
\usage{
methylation_GSE40279
}
\description{
Simulated amount of methylated reads for 650 individuals from GSE40279. The data was sub-sampled to simulate a 30X coverage.
Only 241 sites that are known to differ substaintally between cell types are recorded.
}
\keyword{datasets}
