% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference_blood}
\alias{reference_blood}
\title{A reference containing methylation proportions for pure cell types in the 241 chosen sites.}
\format{A data frame with 241 rows and 7 variables}
\source{
\url{https://bmcbioinformatics.biomedcentral.com/articles/10.1186/s12859-016-0943-7}
}
\usage{
reference_blood
}
\description{
A reference containing methylation proportions for pure cell types in the 241 chosen sites.
}
\keyword{datasets}
